refactor: Analyzer ABC を abc_base.py に分離し循環インポートを完全解消
## 変更概要
循環インポートの根本的な解決のため、Analyzer 抽象基底クラスを独立したモジュールに分離しました。

## 主な変更
### 新規ファイル
- `src/core/analyzer/abc_base.py`: Analyzer ABC の専用モジュール
- `src/core/analyzer/protocols.py`: 将来の型ヒント用プロトコル定義（オプション）

### 修正ファイル
- `src/core/analyzer/base.py`: Analyzer ABC を abc_base からインポート
- `src/core/analyzer/type_inferrer.py`: abc_base.Analyzer を継承
- `src/core/analyzer/dependency_extractor.py`: abc_base.Analyzer を継承

## 解決した問題
### 循環インポート（4パターン）
1. `base.py` → `strategies.py` → `type_inferrer.py` → `base.py`
2. `base.py` → `dependency_extractor.py` → `base.py`
3. `base.py` → `type_inferrer.py` → `base.py`
4. `base.py` → `dependency_extractor.py` → `type_inferrer.py` → `base.py`

### 依存関係の一方向化
```
abc_base.py (基底)
  ↑
  ├─ type_inferrer.py
  ├─ dependency_extractor.py
  └─ base.py (FullAnalyzer, ファクトリ関数)
       ↑
       └─ strategies.py
```

## 検証結果
- ✅ Pyright: エラー0件、循環インポート警告0件（修正前: 4件）
- ✅ mypy: エラーなし（44ファイル）
- ✅ pytest: 279テスト通過、3スキップ
- ✅ カバレッジ: 79%維持

## 関連Issue
Closes #10
